Registration now checks for whether the username hase been used. Streamlined a test
