+++ Model/Enums/RegistrationResults.cs
@@ -4,6 +4,7 @@ namespace Model.Enums
     public enum RegistrationResults
     {
         EmailTaken,
+        UsernameTaken,
         PasswordNoMatch,
         Succeeded,
         PasswordNotStrongEnough,
